LPD-86301 Move FIPSModeUtil from portal-kernel to portal-security-ldap-api
