--- portal-kernel/src/com/liferay/portal/kernel/security/fips/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
